requirements.txt
aiogram==3.12.1
aiohttp>=3.8.0
python-dotenv>=1.0.0
